Extract shared inventory import engine
